+++ app/src/main/java/org/oxycblt/auxio/playback/replaygain/ReplayGain.kt
@@ -43,3 +43,11 @@ enum class ReplayGainMode {
         }
     }
 }
+
+/** Represents the ReplayGain pre-amp values. */
+data class ReplayGainPreAmp(
+    /** The value to use when ReplayGain tags are present. */
+    val with: Float,
+    /** The value to use when ReplayGain tags are not present. */
+    val without: Float,
+)